Separation of VisitorList class from dashboard class
