+++ NodeSrc/Managers/DashboardManager.js
@@ -15,69 +15,8 @@ class DashboardManager {
 
         this.router = express.Router();
 
-
-        this.router.post("/visitorlist",(req, res) => { this.getAllVisitors(req,res); });
         this.router.post("/metrics",(req, res) => { this.getDashboardMetrics(req,res); });
     }
-
-  	/*
-  	 * @desc Returns all visitors of the app
-  	 */
-  	getAllVisitors(req,res)
-  	{
-        var appid = req.body.appid;
-        var skipIndex = req.body.skipindex;
-        var pageLimit = req.body.pagelimit;
-        var filterId = req.body.filterid;
-
-        var filter = global.db.collection('filters').findOne(
-
-            {_id:filterId},
-
-            function(err,filter)
-            {
-                var filterQuery = JSON.parse(filter.mongoFilter);
-
-                var visitorCollection = global.db.collection('visitors').aggregate([
-                    { $match :
-                        { "$and": [
-                            {
-                              appid:appid
-                            },
-                            filterQuery
-                          ]
-                        }
-                    },
-                    { $sort :
-                        { "visitormetainfo.lastseen" : -1 }
-                    },
-                    { $skip : skipIndex },
-                    { $limit : pageLimit },
-                    {
-                      $lookup:
-                        {
-                          from: "sessions",
-                          localField: "_id",
-                          foreignField: "visitorid",
-                          as: "sessions"
-                        }
-                    }
-                ]).toArray(function(err,visitors)
-                    {
-                        if(err)
-                        {
-                            res.status(500);
-                            return res.send({status:'failure'});
-                        }
-
-                        return res.send({status:visitors});
-                    }
-                );
-            }
-        );
-
-
-  	}
 
   	/*
   	 * @desc Returns all visitors of the app
@@ -172,44 +111,6 @@ class DashboardManager {
                     });
             }
         );
-  	}
-
-  	/*
-  	 * @desc Return data of visitor by ID
-  	 */
-  	getVisitorById(visitorId,callback)
-  	{
-        if(visitorId)
-    	{
-
-            var visitorCollection = global.db.collection('visitors').aggregate([
-                { $match :
-                    { _id: visitorId }
-                },
-                {
-                  $lookup:
-                    {
-                      from: "sessions",
-                      localField: "_id",
-                      foreignField: "visitorid",
-                      as: "sessions"
-                    }
-                }
-            ]).toArray(function(err,visitor)
-                {
-                    if(err)
-                    {
-                        return callback(null);
-                    }
-                    else
-                    {
-                        return callback(visitor);
-                    }
-                }
-            );
-
-    	}
-
   	}
 }
