+++ WebContent/ui/commonscripts.ejs
@@ -22,6 +22,7 @@
 
 <script type="text/javascript" src="/js/src/internal/app.js"></script>
 <script type="text/javascript" src="/js/src/internal/dashboard.js"></script>
+<script type="text/javascript" src="/js/src/internal/visitorlist.js"></script>
 <script type="text/javascript" src="/js/src/internal/user.js"></script>
 <script type="text/javascript" src="/js/src/internal/filter.js"></script>
 <script type="text/javascript" src="/js/src/internal/main.js"></script>
